+++ src/Filters/FilterSet.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the FilterSet class.
+ *
+ * @copyright   Copyright (c) 2023 Vanilo UG
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-01-30
+ *
+ */
+
+namespace VaniloCloud\Filters;
+
+class FilterSet
+{
+}